refactor(components): Rename badge folder to Badge for consistency
- Renamed `packages/components/badge/badge/index.scss` to `packages/components/badge/Badge/index.scss`
- Renamed `packages/components/badge/badge/index.tsx` to `packages/components/badge/Badge/index.tsx`
- Updated export paths in `packages/components/badge/index.ts` to reflect the new folder name
