+++ packages/components/badge/Badge/index.scss
