+++ packages/components/badge/index.ts
@@ -1,2 +1,2 @@
-export { default as Badge } from "./badge";
-export type { BadgeProps } from "./badge";
+export { default as Badge } from "./Badge";
+export type { BadgeProps } from "./Badge";
